Start Using Traits
Begin using traits to eradicate extraneous custom component uses. Reason: When a block has a custom component, it automatically makes that block interactable in adventure mode making the selection box visible in adventure mode and calling a different event in scripts when right clicking on the block, making it difficult to work with.