Fix computer animation
By changing "num_mip_levels" from the default 4 to 1 (the highest possible value that still works, 0 is lower which works but the grainy effect is even worse) computer screensaver animations work much better. This does come at the cost of looking somewhat grainy when looking at grass blocks from far away, but when the user stays within the confines of the map it seems to be non-noticeable.