Update level elevators
Since the world has moved up, there is space to build the level elevators lower and hide the base further. Now, inside the elevator, the bottom of the elevator shaft can no longer be seen. Also, replace the server blocks that used to be used as the base of the elevator with the server blocks with no screen.